+++ pyhearts/_legacy2025/processing/bounds.py
@@ -80,3 +80,4 @@ def calc_bounds(
             upper_std = lower_std + 1
 
     return [center_lower, lower_height, lower_std], [center_upper, upper_height, upper_std]
+